final_project/requirements.txt
pytest
selenium
pytest-xdist
allure-pytest
webdriver-manager
requests
Faker
PyMySQL
SQLAlchemy
Flask
